moved formatting to separate file
